Add class Buffer
